move: 패키지 이동
- common/application/service -> common/application/util
