+++ loan-mate/src/main/java/com/fisa/bank/common/application/util/ai/AiClient.java
@@ -1,4 +1,4 @@
-package com.fisa.bank.common.application.service;
+package com.fisa.bank.common.application.util.ai;
 
 public interface AiClient {
   <T, B> T fetchOne(String endpoint, B body, Class<T> clazz);
