+++ loan-mate/src/main/java/com/fisa/bank/common/application/util/ai/AiClientImpl.java
@@ -1,4 +1,4 @@
-package com.fisa.bank.common.application.service;
+package com.fisa.bank.common.application.util.ai;
 
 import lombok.RequiredArgsConstructor;
