+++ loan-mate/src/main/java/com/fisa/bank/common/application/util/jwt/DevAccessTokenManager.java
@@ -1,4 +1,4 @@
-package com.fisa.bank.common.application.service;
+package com.fisa.bank.common.application.util.jwt;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
